// src/client/app/selection/shared/CNCMachine/model/product-condition.ts
export class ProductCondition {
    maxFeedSpeed: number;
    tableTravel: number;
    productMaxMass: number;
    tableMass: number;
    productMaxHeight: number;
    loadCharacter: string;
    productMaxLength: number;
    feedAcceleration: number;
    productMaxWidth: number;
    spindleBoxMass: number;
    tableLength: number;
    productStiffness: number;
}